+++ Scripts/Attributes/Editor/VerticalGroupDrawer.cs
@@ -1,0 +1,31 @@
+using UnityEditor;
+using UnityEngine;
+
+namespace EditorAttributes.Editor
+{
+    [CustomPropertyDrawer(typeof(VerticalGroupAttribute))]
+    public class VerticalGroupDrawer : PropertyDrawerBase
+    {
+		public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
+		{
+			var verticalGroup = attribute as VerticalGroupAttribute;
+			var serializedObject = property.serializedObject;
+		
+			EditorGUILayout.BeginVertical();
+
+			EditorGUIUtility.labelWidth = verticalGroup.LabelWidth;
+			EditorGUIUtility.fieldWidth = verticalGroup.FieldWidth;
+		
+			foreach (string variableName in verticalGroup.FieldsToGroup)
+			{
+				var variableProperty = serializedObject.FindProperty(variableName);
+
+				if (variableProperty != null) EditorGUILayout.PropertyField(variableProperty, true);
+			}
+
+			EditorGUILayout.EndVertical();
+		}
+
+		public override float GetPropertyHeight(SerializedProperty property, GUIContent label) => -EditorGUIUtility.standardVerticalSpacing; // Remove the space for the hidden field
+	}
+}