+++ Scripts/Attributes/TagDropdownAttribute.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+namespace EditorAttributes
+{
+    public class TagDropdownAttribute : PropertyAttribute { }
+}