+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+EditorAttributes v1.4.0 Changelog:
+- Added Prefix/Sufix attribute
+- Added DrawLine attribute
+- Added TagDropdown attribute
+- Added Image attribute
+- Added VerticalGroup attribute
+- Added SceneDropdown attribute
+
 EditorAttributes v1.3.0 Changelog:
 - Added MinMaxSlider attribute
 - Added Clamp attribute